Add option to hydrate synchronously
The original behavior to hydrate asynchronously causes the DOM subtree
corresponding to the MDX output to be entirely replaced by React
reconciliation once async hydration is complete (which switches from a
div with dangerouslySetInnerHTML to an MDXProvider component). This
works okay when the Markdown contains only simple elements like
paragraphs with text, but it's jarring in other settings, e.g. when
there are videos in the document.

This patch adds an option to render synchronously, so there's no
switch-over from the div to the MDXProvider, so there's no extraneous
DOM updates that happen.

See https://github.com/hashicorp/next-mdx-remote/issues/118 for
additional context.